fix: changed DTO class for stat

+++ stat-service/src/main/java/ru/yandex/practicum/stats/dto/ViewStats.java
@@ -11,10 +11,9 @@ import lombok.Setter;
 @NoArgsConstructor
 public class ViewStats {
 
-
     private String app;
 
     private String uri;
 
-    private Integer hits;
+    private Long hits;
 }
